tweaks to user admin styling
