+++ vuu-ui/sample-apps/feature-user-admin/src/UserAdmin.css
@@ -1,29 +1,10 @@
 .vuuUserAdmin {
   display: flex;
   flex-direction: column;
-  height: 100vh;
+  height: 100%;
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif;
   background-color: #f5f5f5;
   color: #333;
-}
-
-.vuuUserAdmin-header {
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  color: white;
-  padding: 24px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.vuuUserAdmin-header h1 {
-  margin: 0 0 8px 0;
-  font-size: 28px;
-  font-weight: 600;
-}
-
-.vuuUserAdmin-header p {
-  margin: 0;
-  font-size: 14px;
-  opacity: 0.9;
 }
 
 .vuuUserAdmin-tabs {
@@ -32,30 +13,6 @@
   background: white;
   border-bottom: 1px solid #e0e0e0;
   box-shadow: 0 1px 3px rgba(0, 0, 0, 0.05);
-}
-
-.tab {
-  flex: 1;
-  padding: 16px 24px;
-  border: none;
-  background: none;
-  cursor: pointer;
-  font-size: 14px;
-  font-weight: 500;
-  color: #666;
-  border-bottom: 3px solid transparent;
-  transition: all 0.3s ease;
-}
-
-.tab:hover {
-  background-color: #fafafa;
-  color: #333;
-}
-
-.tab.active {
-  color: #667eea;
-  border-bottom-color: #667eea;
-  background-color: #f9f9fb;
 }
 
 .vuuUserAdmin-content {
@@ -143,34 +100,6 @@
 .roles-table tbody tr:hover,
 .groups-table tbody tr:hover {
   background-color: #fafafa;
-}
-
-.users-table td,
-.roles-table td,
-.groups-table td {
-  padding: 12px;
-}
-
-.users-table button,
-.roles-table button,
-.groups-table button {
-  padding: 6px 12px;
-  margin-left: 4px;
-  background-color: #f0f0f0;
-  color: #333;
-  border: 1px solid #d0d0d0;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 12px;
-  font-weight: 500;
-  transition: all 0.2s ease;
-}
-
-.users-table button:hover,
-.roles-table button:hover,
-.groups-table button:hover {
-  background-color: #e0e0e0;
-  border-color: #b0b0b0;
 }
 
 .user-details-section {
@@ -647,13 +576,10 @@
   gap: 10px;
   height: 100%;
   min-height: 0;
-  padding: 10px;
   width: 560px;
 }
 
 .vuuUserAdmin-drawerSection {
-  background: #ffffff;
-  border: 1px solid #e0e0e0;
   border-radius: 6px;
   display: flex;
   flex: 1;
